V0.3

+++ lib/app/modules/pagamento/modules/pagamento_resumo.dart
@@ -1,6 +1,6 @@
 class PagamentoResumo {
   String data;
-  int valor;
+  double valor;
   String referencia;
   String nomeFicheiroMeps;
   String dataInsercaoMensagem;
